Flask/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
future==0.18.3
gevent==23.9.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pycparser==2.21
PySocks==1.7.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
tldextract==5.1.1
urllib3==2.1.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
